# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ChiefVoice CRM API
  version: 1.0.0
  description: |
    AI-powered voice CRM backend.

    **Authentication:** All protected endpoints require a identity-provider-issued ID token.
    Pass it as `Authorization: Bearer <token>`.
    The SSE endpoint `/api/logs-stream` uses a short-lived ticket created by `/api/logs-stream/ticket`
    because browser EventSource cannot set custom Authorization headers. Other
    protected endpoints do not accept JWTs in query strings.

    **Roles:** `Organization Admin`, `Super Admin`, `Agent`, `Manager`.
    Platform-admin endpoints additionally require the email to be in `PLATFORM_ADMIN_EMAILS`.

  contact:
    name: ChiefVoice Support
    email: support@chiefvoice.ai
servers:
  - url: /api
    description: Current server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication ID token
...
  schemas:
...
    VirtualNumber:
      type: object
      properties:
        id:       { type: string }
        number:   { type: string }
        provider: { type: string }
        label:    { type: string }
    TeamMember:
      type: object
      properties:
        id:    { type: string }
        name:  { type: string }
        email: { type: string }
        role:  { type: string }
...
security:
  - BearerAuth: []

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /settings/numbers:
    get:
      tags: [Settings]
      summary: List virtual numbers
      responses:
        '200':
          description: Numbers
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/VirtualNumber' } }
    post:
      tags: [Settings]
      summary: Add virtual number
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VirtualNumber' }
      responses:
        '201': { description: Created }
  /settings/numbers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Settings]
      summary: Update number
      responses:
        '200': { description: Updated }
    delete:
      tags: [Settings]
      summary: Delete number
      responses:
        '200': { description: Deleted }
...
  /settings/team:
    get:
      tags: [Settings]
      summary: List team members
      responses:
        '200':
          description: Team
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/TeamMember' } }
    post:
      tags: [Settings]
      summary: Add team member
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeamMember' }
      responses:
        '201': { description: Created }
  /settings/team/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Settings]
      summary: Update team member
      responses:
        '200': { description: Updated }
...
  /settings/org:
    get:
      tags: [Settings]
      summary: Get org settings
      responses:
        '200': { description: Org settings }
    post:
      tags: [Settings]
      summary: Update org settings
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Saved }

  # ── Knowledge ─────────────────────────────────────────────────────────────
...